[REF] Split dir/file utilities
