+++ images/common/spank-nccl-debug/CMakeLists.txt
@@ -8,6 +8,7 @@ set(LIBRARY_NAME nccldebug)
 set(SOURCES
         ${CMAKE_CURRENT_SOURCE_DIR}/src/snccld.c
         ${CMAKE_CURRENT_SOURCE_DIR}/src/snccld_state.c
+        ${CMAKE_CURRENT_SOURCE_DIR}/src/snccld_util_dir_file.c
         ${CMAKE_CURRENT_SOURCE_DIR}/src/snccld_util_host.c
         ${CMAKE_CURRENT_SOURCE_DIR}/src/snccld_util_oplock.c
 )
