+++ images/common/spank-nccl-debug/src/snccld_state.c
@@ -1,5 +1,7 @@
 #include "snccld_state.h"
+
 #include "snccld.h"
+#include "snccld_util_dir_file.h"
 
 #include <fcntl.h>
 #include <limits.h>
